Review feedback - rename components for shorter name
