+++ airflow/ui/src/components/FlexibleForm/FieldString.tsx
@@ -16,37 +16,30 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-import { json } from "@codemirror/lang-json";
-import { githubLight, githubDark } from "@uiw/codemirror-themes-all";
-import CodeMirror from "@uiw/react-codemirror";
-
-import { useColorMode } from "src/context/colorMode";
+import { Input } from "@chakra-ui/react";
 
 import type { FlexibleFormElementProps } from ".";
 
-export const FlexibleFormFieldObject = ({ name, param }: FlexibleFormElementProps) => {
-  const { colorMode } = useColorMode();
-
-  return (
-    <CodeMirror
-      basicSetup={{
-        autocompletion: true,
-        bracketMatching: true,
-        foldGutter: true,
-        lineNumbers: true,
-      }}
-      extensions={[json()]}
-      height="200px"
+export const FieldString = ({ name, param }: FlexibleFormElementProps) => (
+  <>
+    <Input
+      defaultValue={String(param.value)}
       id={`element_${name}`}
-      style={{
-        border: "1px solid #CBD5E0",
-        borderRadius: "8px",
-        outline: "none",
-        padding: "2px",
-        width: "100%",
-      }}
-      theme={colorMode === "dark" ? githubDark : githubLight}
-      value={JSON.stringify(param.value, undefined, 2)}
+      list={param.schema.examples ? `list_${name}` : undefined}
+      maxLength={param.schema.maxLength ?? undefined}
+      minLength={param.schema.minLength ?? undefined}
+      name={`element_${name}`}
+      placeholder={param.schema.examples ? "Start typing to see proposal values." : undefined}
+      size="sm"
     />
-  );
-};
+    {param.schema.examples ? (
+      <datalist id={`list_${name}`}>
+        {param.schema.examples.map((example) => (
+          <option key={example} value={example}>
+            {example}
+          </option>
+        ))}
+      </datalist>
+    ) : undefined}
+  </>
+);
